requirements.txt
pandas>=2.2.0
pyyaml>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
datasets>=2.20.0
scikit-learn>=1.4.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0
kagglehub>=0.2.5
streamlit
nbformat
python-docx
